+++ CHANGELOG.md
@@ -4,6 +4,9 @@ on [Keep a CHANGELOG](http://keepachangelog.com/). This project adheres to
 [Semantic Versioning](http://semver.org/).
 
 ### [Unreleased]
+### Changed
+- ReplaySingleEvent JMX commands can now take an optional commandRuntimeString of the event source name
+  to work with MI contexts
 
 # [17.104.0-M10] - 2025-11-03
 ### Added